+++ onedrive-api-test/src/main.rs
@@ -1,5 +1,5 @@
 use anyhow::{ensure, Context as _, Result};
-use onedrive_api::{Auth, Permission, Tenant};
+use onedrive_api::{Auth, ClientCredential, Permission, Tenant};
 use std::{
     env,
     fs::File,
@@ -12,7 +12,6 @@ const DEFAULT_REDIRECT_URI: &str = "https://login.microsoftonline.com/common/oau
 #[derive(Debug)]
 struct Args {
     client_id: String,
-    client_secret: Option<String>,
     output_file: String,
     redirect_uri: String,
 }
@@ -25,14 +24,12 @@ fn parse_args() -> Result<Args> {
     let redirect_uri = args
         .opt_value_from_str("-r")?
         .unwrap_or_else(|| DEFAULT_REDIRECT_URI.to_owned());
-    let client_secret = args.opt_value_from_str("-s")?;
     let client_id = args
         .free_from_str::<String>()
         .context("Missing client id")?;
     ensure!(args.finish().is_empty(), "Too many arguments");
     Ok(Args {
         client_id,
-        client_secret,
         output_file,
         redirect_uri,
     })
@@ -43,7 +40,7 @@ fn exit_with_help() -> ! {
         "
 This binary helps you to get an authorized refresh token to be used in tests.
 
-USAGE: {} [-o <output_file>] [-r <redirect_uri>] [-s <client_secret>] <client_id>
+USAGE: {} [-o <output_file>] [-r <redirect_uri>] <client_id>
     `output_file` is default to be `.env`
     `redirect_url` is default to be `{}`
 ",
@@ -69,7 +66,7 @@ async fn main() -> Result<()> {
     );
     let url = auth.code_auth_url();
     eprintln!("Code auth url: {url}");
-    if open::that(url).is_err() {
+    if open::that(url.as_str()).is_err() {
         eprintln!("Cannot open browser, please open the url above manually.");
     }
     eprintln!("Please login in browser, paste the redirected URL here and then press <Enter>");
@@ -90,9 +87,7 @@ async fn main() -> Result<()> {
     };
 
     eprintln!("Logining...");
-    let token = auth
-        .login_with_code(&code, args.client_secret.as_deref())
-        .await?;
+    let token = auth.login_with_code(&code, &ClientCredential::None).await?;
     let refresh_token = token.refresh_token.expect("Missing refresh token");
 
     {
@@ -100,9 +95,6 @@ async fn main() -> Result<()> {
         writeln!(f, "export ONEDRIVE_API_TEST_CLIENT_ID='{}'", args.client_id)?;
         writeln!(f, "export ONEDRIVE_API_TEST_REDIRECT_URI='{}'", args.redirect_uri)?;
         writeln!(f, "export ONEDRIVE_API_TEST_REFRESH_TOKEN='{refresh_token}'")?;
-        if let Some(client_secret) = args.client_secret {
-            writeln!(f, "export ONEDRIVE_API_TEST_CLIENT_SECRET='{client_secret}'")?;
-        }
     }
 
     Ok(())